+++ src/Components/Alert/PopUp/CreateSample/FooterButtons.tsx
@@ -58,5 +58,5 @@ export const FooterButtons = memo((props: {
         }}
       />
     </View>
-  )
+  );
 });