+++ src/Globals/AppRoutes.ts
@@ -10,7 +10,7 @@ export default class AppRoutes {
   static SS_THEME_SCOPE               = `${this.SETTINGS_SCOPE}/ThemeScope`;
   static SS_CREDENTIALS_SCOPE         = `${this.SETTINGS_SCOPE}/CredentialScope`;
 
-  static PROJECT_SCOPE   = (id_project: string) => `/${id_project}`;
-  static PS_SAMPLE_SCOPE = (id_project: string, id_sample: string) => `${this.PROJECT_SCOPE(id_project)}/${id_sample}`;
-
+  static PROJECT_SCOPE           = (id_project: string) => `/${id_project}`;
+  static PS_EXPORT_PORJECT_SCOPE = (id_project: string) => `${this.PROJECT_SCOPE(id_project)}/exportProjectScope`;
+  static PS_SAMPLE_SCOPE         = (id_project: string, id_sample: string) => `${this.PROJECT_SCOPE(id_project)}/${id_sample}`;
 }