+++ src/Screens/ExportProject/__LC__/index.tsx
@@ -1,0 +1,5 @@
+import { AvailableExportFormatButtons } from './SettingsButtons';
+
+export const LC = {
+  AvailableExportFormatButtons: AvailableExportFormatButtons,
+};