+++ src/Services/UtilService.ts
@@ -10,6 +10,7 @@ export default class UtilService {
     'noSpaces':        /\s/,
     'noSpecialLetter': /^[a-zA-Z0-9]+$/,
     'id':              /^[0-9A-Za-z-]+$/,
+    'fileName':        /^[a-zA-Z0-9_-]+$/,
     'hexColor':        /^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$/,
   };