+++ src/Translations/Screen/ExportProjectScreen.ts
@@ -1,0 +1,14 @@
+import { LanguageTag } from '@Types/AppTypes';
+
+type TranslationDTO = Record<LanguageTag, {
+  'Docx': string
+}>
+
+export const R_Screen_ExportProject: TranslationDTO = {
+  'en-US': {
+    'Docx': 'Docx',
+  },
+  'pt-BR': {
+    'Docx': 'Docx',
+  },
+};